chore: split user model into user-login and user-info; add associations between user-login and user-info
