+++ server/db/models/index.js
@@ -1,4 +1,5 @@
-const User = require('./user')
+const UserLogin = require('./user-login')
+const UserInfo = require('./user-info')
 
 /**
  * If we had any associations to make, this would be a great place to put them!
@@ -7,6 +8,9 @@ const User = require('./user')
  *    BlogPost.belongsTo(User)
  */
 
+UserLogin.hasOne(UserInfo)
+UserInfo.belongsTo(UserLogin)
+
 /**
  * We'll export all of our models here, so that any time a module needs a model,
  * we can just require it from 'db/models'
@@ -14,5 +18,6 @@ const User = require('./user')
  * instead of: const User = require('../db/models/user')
  */
 module.exports = {
-  User
+  UserLogin,
+  UserInfo
 }
