--- server/db/models/user.js
@@ -1,134 +1,0 @@
-const crypto = require('crypto')
-const Sequelize = require('sequelize')
-const db = require('../db')
-
-const User = db.define('user', {
-  firstName: {
-    type: Sequelize.STRING,
-    allowNull: false
-  },
-  lastName: {
-    type: Sequelize.STRING,
-    allowNull: false
-  },
-  username: {
-    type: Sequelize.STRING,
-    unique: true,
-    allowNull: false
-  },
-  email: {
-    type: Sequelize.STRING,
-    unique: true,
-    allowNull: false
-  },
-  password: {
-    type: Sequelize.STRING,
-    // Making `.password` act like a func hides it when serializing to JSON.
-    // This is a hack to get around Sequelize's lack of a "private" option.
-    get() {
-      return () => this.getDataValue('password')
-    }
-  },
-  salt: {
-    type: Sequelize.STRING,
-    // Making `.salt` act like a function hides it when serializing to JSON.
-    // This is a hack to get around Sequelize's lack of a "private" option.
-    get() {
-      return () => this.getDataValue('salt')
-    }
-  },
-  googleId: {
-    type: Sequelize.STRING
-  },
-  billingAddress: {
-    type: Sequelize.STRING,
-    allowNull: false
-  },
-  shippingAddress: {
-    type: Sequelize.STRING,
-    allowNull: false
-  },
-  phoneNumber: {
-    type: Sequelize.STRING,
-    allowNull: false
-  },
-  CCInfo: {
-    type: Sequelize.STRING,
-    validate: {
-      isCreditCard: true
-    },
-    get() {
-      return () => this.getDataValue('CCInfo')
-    }
-  },
-  DOB: {
-    type: Sequelize.DATEONLY,
-    allowNull: false
-  },
-  imageUrl: {
-    type: Sequelize.TEXT,
-    validate: {
-      isUrl: true
-    }
-  }
-})
-
-module.exports = User
-
-/**
- * instanceMethods
- */
-User.prototype.correctPassword = function(candidatePwd) {
-  return User.encryptPassword(candidatePwd, this.salt()) === this.password()
-}
-
-/**
- * classMethods
- */
-User.generateSalt = function() {
-  return crypto.randomBytes(16).toString('base64')
-}
-
-User.encryptPassword = function(plainText, salt) {
-  return crypto
-    .createHash('RSA-SHA256')
-    .update(plainText)
-    .update(salt)
-    .digest('hex')
-}
-
-User.encryptCCInfo = function(plainText, salt) {
-  return crypto
-    .createHash('RSA-SHA256')
-    .update(plainText)
-    .update(salt)
-    .digest('hex')
-}
-
-/**
- * hooks
- */
-const setSaltAndPassword = user => {
-  if (user.changed('password')) {
-    user.salt = User.generateSalt()
-    user.password = User.encryptPassword(user.password(), user.salt())
-  }
-}
-
-const setSaltAndCCInfo = user => {
-  if (user.changed('CCInfo')) {
-    user.salt = User.generateSalt()
-    user.CCInfo = User.encryptCCInfo(user.CCInfo(), user.salt())
-  }
-}
-
-User.beforeCreate(setSaltAndPassword)
-User.beforeUpdate(setSaltAndPassword)
-User.beforeBulkCreate(users => {
-  users.forEach(setSaltAndPassword)
-})
-User.beforeCreate(setSaltAndCCInfo)
-User.beforeUpdate(setSaltAndCCInfo)
-User.beforeBulkCreate(users => {
-  users.forEach(setSaltAndCCInfo)
-})
